ci fix & review comments
