+++ openmetadata-spec/src/main/resources/json/schema/entity/services/connections/dashboard/supersetConnection.json
@@ -20,7 +20,8 @@
           "title": "Provider",
           "description": "Authentication provider for the Superset service. For basic user/password authentication, the default value `db` can be used. This parameter is used internally to connect to Superset's REST API.",
           "type": "string",
-          "default": "db"
+          "default": "db",
+          "enum": ["db","ldap"]
         },
         "username": {
           "title": "Username",
